config/requirements.txt
absl-py==2.4.0
astunparse==1.6.3
audioread==3.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
flatbuffers==25.12.19
fonttools==4.62.1
gast==0.7.0
google-pasta==0.2.0
grpcio==1.80.0
h5py==3.14.0
idna==3.11
joblib==1.5.3
keras==3.12.1
kiwisolver==1.5.0
lazy-loader==0.5
libclang==18.1.1
librosa==0.11.0
llvmlite==0.47.0
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
msgpack==1.1.2
namex==0.1.0
numba==0.65.0
numpy==2.2.6
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==2.3.3
pillow==12.2.0
platformdirs==4.9.4
pooch==1.9.0
protobuf==7.34.1
pycparser==3.0
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.33.1
rich==14.3.3
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
wrapt==2.1.2
